Use correct function for determining size of C struct.
Previously, sys.getsizeof() was used for calculating the size of a
ctypes.Struct before receiving data from libc into it. The manual for
getsizeof(obj) specifies that it returns the size of "obj" _plus data for
garbage collection, if any_. On my 64-bit Ubuntu 15.04, getsizeof() returns 144
whereas ctypes.sizeof() for the same struct returns 16, which is the sought
size.